Move view decorators into new LTIHService
Move a bunch of view decorators (upsert_h_user, upsert_h_group,
add_user_to_group, ...) into a new LTIHService and have the views call
the methods on this service rather than using them as view decorators.

This means that upsert_h_user etc are now called *after* request
validation rather than before (the view's decorators are called before
the view's validation schema). So if the request is invalid we'll return
a ValidationError before upsert_h_user etc get called, rather than
potentially calling upsert_h_user etc with an invalid request.

Calling upsert_h_user etc with an invalid request could cause problems
because those view decorators run code that uses the request params. In
particular they make calls to LTILaunchResource methods that use the
request params. This change ensures that the request params are always
valid when this happens.
