+++ lms/views/content_item_selection.py
@@ -37,18 +37,20 @@ https://canvas.instructure.com/doc/api/file.content_item.html
 from pyramid.view import view_config
 
 from lms.views import helpers
-from lms.views.decorators import upsert_course_group, upsert_h_user
 
 
 @view_config(
     authorized_to_configure_assignments=True,
-    decorator=[upsert_h_user, upsert_course_group],
     permission="launch_lti_assignment",
     renderer="lms:templates/file_picker.html.jinja2",
     request_method="POST",
     route_name="content_item_selection",
 )
 def content_item_selection(context, request):
+    lti_h_service = request.find_service(name="lti_h")
+    lti_h_service.upsert_h_user()
+    lti_h_service.upsert_course_group()
+
     context.js_config.update(
         {
             # The URL that the JavaScript code will open if it needs the user to
