+++ lms/views/decorators/__init__.py
@@ -18,18 +18,10 @@ For documentation see:
 
 https://docs.pylonsproject.org/projects/pyramid/en/latest/narr/viewconfig.html#non-predicate-arguments
 """
-from lms.views.decorators.h_api import (
-    add_user_to_group,
-    upsert_course_group,
-    upsert_h_user,
-)
 from lms.views.decorators.lis_result_sourcedid import upsert_lis_result_sourcedid
 from lms.views.decorators.reports import report_lti_launch
 
 __all__ = (
-    "upsert_h_user",
-    "upsert_course_group",
-    "add_user_to_group",
     "report_lti_launch",
     "upsert_lis_result_sourcedid",
 )
