+++ tox.ini
@@ -44,7 +44,7 @@ passenv =
     dev: VIA_URL
 deps =
     {tests,clean,coverage}: coverage
-    {tests,lint}: httpretty
+    {tests,functests,lint}: httpretty
     {tests,functests,lint,docstrings,checkdocstrings}: pytest
     {tests,functests,lint,docstrings,checkdocstrings}: factory-boy
     {tests,functests,lint,docstrings,checkdocstrings}: h-matchers>=1.2.0
